+++ source/scriptlib/ScriptVariant.h
@@ -22,7 +22,7 @@ typedef enum VariantType {
 typedef struct ScriptVariant {
 	union			//value
 	{
-		LONG lVal;
+		s32 lVal;
 		VOID *ptrVal;
 		DOUBLE dblVal;
 		int strVal;
@@ -48,7 +48,7 @@ void ScriptVariant_Init(ScriptVariant * var);
 void ScriptVariant_Copy(ScriptVariant * svar, ScriptVariant * rightChild);	// faster in some situations
 void ScriptVariant_ChangeType(ScriptVariant * var, VARTYPE cvt);
 VARTYPE ScriptVariant_GetType(ScriptVariant * var);
-s32 ScriptVariant_IntegerValue(ScriptVariant * var, LONG * pVal);
+s32 ScriptVariant_IntegerValue(ScriptVariant * var, s32 * pVal);
 s32 ScriptVariant_DecimalValue(ScriptVariant * var, DOUBLE * pVal);
 BOOL ScriptVariant_IsTrue(ScriptVariant * svar);
 void ScriptVariant_ToString(ScriptVariant * svar, char* buffer);
